Remove misleading quote
Out of context, you might think that I misspoke and was actually talking about WMF when I said "Wikipedia". But I was referring to the toxicity of the English Wikipedia editor community, which is the main hypothesis on the cause of the editor decline on the English Wikipedia. My point was that Lila can't claim credit for fixing the editor decline when the cause of it still remains, and no action has been taken.